+++ superset-frontend/src/components/index.ts
@@ -1,1 +1,1 @@
-export { default as Select } from './Select/Select';
+export { default as Select } from './Select/AntdSelect';
